wip; part2 soluion is uint

+++ src/Day17/Services/Instructions/Adv.cs
@@ -11,7 +11,7 @@ public static class Adv
 {
     public static int Opcode { get; set; } = 0;
 
-    public static int Execute(int? comboOperand, int registerAValue)
+    public static uint Execute(uint? comboOperand, uint registerAValue)
     {
         if (comboOperand == null)
         {
@@ -19,8 +19,8 @@ public static class Adv
         }
 
         var numerator = registerAValue;
-        var denominator = Math.Pow(2, (int)comboOperand);
+        var denominator = Math.Pow(2, (uint)comboOperand);
 
-        return (int)(numerator / denominator);
+        return (uint)(int)(numerator / denominator);
     }
 }
